refactor: popover에 대해 라이브러리 사용해서 리팩토링
